trie: widen NodeId to 64-bit (+0.34 %)
NodeId is a blob offset declared `enum class : uint32_t`, so the RISC-V 32-bit forms (addw, addiw,
srliw, subw) were emitted wherever it is arithmetic — an instruction class the reth guest never
uses. Eight casts.

The wire format is deliberately untouched: append_u32 and rd_u32 still read and write four bytes,
because there the width is part of the encoding rather than an implementation choice.

Measured +0.34 % of the guest's own work, 16/16 state roots PASS.

Worth recording for method: the STATIC instruction count moved by six, out of 3,900. The runtime
moved forty times that, because those six sit in the reader's hottest paths (upsert_node,
erase_node, the node matcher, fold_ext_node_path_maybe). Counting instructions in a binary says
what code exists, not how often it runs.

+++ category/execution/ethereum/db/offset_trie_db.cpp
@@ -91,14 +91,14 @@ namespace mpt_witness
 
         // Empty trie: reserve the sentinel root id so fresh_id never reuses it.
         if (is_overlay_id(root)) {
-            next_id_ = NodeId{static_cast<uint32_t>(root) + 1};
+            next_id_ = NodeId{static_cast<uint64_t>(root) + 1};
         }
     }
 
     void TrieStore::prime_node(NodeViewBase const node)
     {
         auto const id = static_cast<NodeId>(
-            static_cast<uint32_t>(node.bytes() - blob_.data()));
+            static_cast<uint64_t>(node.bytes() - blob_.data()));
         if (node.tag() == DIGEST) {
             // A digest's hash IS the 32 bytes in the blob; child_ref reads it there.
             return;
@@ -201,7 +201,7 @@ namespace mpt_witness
         // current reads consult the overlay first.
         NodeViewBase const node = [&]() -> NodeViewBase {
             if constexpr (priming_pass) {
-                MONAD_ASSERT(static_cast<uint32_t>(id) < blob_.size());
+                MONAD_ASSERT(static_cast<uint64_t>(id) < blob_.size());
                 return get_original(id);
             }
             else {
@@ -418,7 +418,7 @@ namespace mpt_witness
     NodeId TrieStore::fresh_id()
     {
         NodeId const fresh = next_id_;
-        next_id_ = NodeId{static_cast<uint32_t>(next_id_) + 1};
+        next_id_ = NodeId{static_cast<uint64_t>(next_id_) + 1};
         return fresh;
     }
